distributed-docker/machine-learging/requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==1.22.4
pandas==1.3.5
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
tzdata==2024.2
Werkzeug==3.1.3
